fix: ensure productivityScore is double for DailyStats to fix flutter build error

+++ lib/features/analytics/providers/time_tracking_provider.dart
@@ -207,7 +207,7 @@ class TimeTrackingProvider extends ChangeNotifier {
       totalTimeSpent: Duration(minutes: totalMinutes.toInt()),
       tasksCompleted: tasksCompleted,
       pomodoroSessions: pomodoroSessions,
-      productivityScore: productivityScore,
+      productivityScore: productivityScore.toDouble(),
     );
   }
